More features
Helper Methods for the following counts:
Hash tags
Negation words
all uppercase words
-input: Each of these take a raw tweet and counts various features.
Preprocessing shouldn't happen here.
-output: A number representing the counts

Part of speech tagging:
tweet_pos_tags
- input: tweet_tokens
-output : returns a dictionary that is the tagged words in the tweet
Takes a list of tokens (preprocessed or not) and returns the various
taggings for the tweet

count_pos_tags
-input: pos_list , normalized
-output: dictionary of final counts normalized or not
normalized = 0 (not normalized)
normalized = 1 (normalized, takes the proportion of tag to the rest of
the sentence tags)
Takes 2 inputs, a part of speech tag list and a normalization variable.
Returns a dictionary with each count value mapped to the POS tag.